+++ changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 1.0.0.5
+
 - Added the command `/ess convert HuskHomes`, allows converting the database from HuskHomes to zEssentials.
 - Added the command `/ess convert AxVaults`, allows converting the database from AxVaults to zEssentials.
 - Added the command `/mailbox give <player> <item> [<amount>]`, Add an item to a player’s mailbox.